Moved `src/enumerations/paths/parts/Scheme.php to `src/enumerations/paths/parts/url/Scheme.php`.
Moved `tests/enumerations/paths/parts/SchemeTest.php to `tests/enumerations/paths/parts/url/SchemeTest.php`.

Updated namespaces appropriately.
